NetBSD warnings removed

+++ ex.c
@@ -254,7 +254,7 @@ main(int ac, char **av)
 		case 'w':
 			defwind = 0;
 			if (av[0][2] == 0) defwind = 3;
-			else for (cp = &av[0][2]; isdigit(*cp); cp++)
+			else for (cp = &av[0][2]; isdigit((int)*cp); cp++)
 				defwind = 10*defwind + *cp - '0';
 			break;
