+++ ex_io.c
@@ -269,12 +269,12 @@ glob(struct glob *gp)
 		register char *v = genbuf + 5;		/* strlen("echo ") */
 
 		for (;;) {
-			while (isspace(*v))
+			while (isspace((int)*v))
 				v++;
 			if (!*v)
 				break;
 			*argv++ = cp;
-			while (*v && !isspace(*v))
+			while (*v && !isspace((int)*v))
 				*cp++ = *v++;
 			*cp++ = 0;
 			gp->argc0++;
