+++ ex_voper.c
@@ -15,7 +15,7 @@ static char sccsid[] = "@(#)ex_voper.c	8.1 (Berkeley) 6/9/93";
 #include "ex_tty.h"
 #include "ex_vis.h"
 
-#define	blank()		isspace(wcursor[0])
+#define	blank()		isspace((int)wcursor[0])
 #define	forbid(a)	if (a) goto errlab;
 
 static int edge(void);
@@ -279,7 +279,7 @@ ein:
 	 */
 	case ',':
 		forbid (lastFKND == 0);
-		c = isupper(lastFKND) ? tolower(lastFKND) : toupper(lastFKND);
+		c = isupper((int)lastFKND) ? tolower((int)lastFKND) : toupper((int)lastFKND);
 		i = lastFCHR;
 		if (vglobp == 0)
 			vglobp = "";
@@ -673,7 +673,7 @@ slerr:
 		if (any(*globp, "^+-."))
 			c = *globp++;
 		i = 0;
-		while (isdigit(*globp))
+		while (isdigit((int)*globp))
 			i = i * 10 + *globp++ - '0';
 		if (any(*globp, "^+-."))
 			c = *globp++;
@@ -826,7 +826,7 @@ int
 wordof(int which, char *wc)
 {
 
-	if (isspace(*wc))
+	if (isspace((int)*wc))
 		return (0);
 	return (!wdkind || wordch(wc) == which);
 }
